+++ planning/__init__.py
@@ -4,4 +4,4 @@ PyPlanning
 ----------
 """
 
-__version__ = "1.5.9"
+__version__ = "1.5.10"