+++ planning/gui/mainwindow.py
@@ -26,7 +26,7 @@ from qtpy.compat import getopenfilename, getsavefilename
 #  Local imports
 from planning import __version__
 from planning.config import APP_DESC, APP_NAME, DEBUG, DEBUG_VAR_STR, Conf, _
-from planning.gantt import init_log_to_sysout
+from planning.gantt import LOG
 from planning.gui.centralwidget import PlanningCentralWidget
 from planning.gui.logviewer import exec_logviewer_dialog
 from planning.utils import qthelpers as qth
@@ -63,7 +63,7 @@ class PlanningMainWindow(QW.QMainWindow):
                 multithreaded=False,
             )
             if DEBUG >= 1:
-                init_log_to_sysout(stream=self.console)
+                LOG.initialize(stream=self.console)
             dockwidget, location = self.console.create_dockwidget("Console")
             self.addDockWidget(location, dockwidget)
             dockwidget.hide()
@@ -493,11 +493,7 @@ Please check the file content."""
 
                     # ------------------------------------------------------------------
                     # TODO: Remove the following code when guidata 3.2.3 is released
-                    import logging  # pylint: disable=C0415
-
-                    logger = logging.getLogger("Gantt")
-                    for handler in logger.handlers:
-                        handler.setStream(None)
+                    LOG.close()
                     # ------------------------------------------------------------------
 
                 except RuntimeError: